alter settings to include default values
This was complicated but I reworked get_usable_kwargs to record
all of the settings, not just the ones passed in via kwargs. This
means that the default values of a function like torchaudio's
MelSpectrogram will be recorded in the settings of the AudioSpectrogram.

This was a necessary change to ensure that transforms like CropTime have
access to values like hop_length for how the sg was generated and are
relevant to interpreting it (hop_length and sample_rate can be used to
tell the duration of a spectrogram, which is used for time cropping)